Bytes: move to internal numbers package.
