+++ internal/numbers/numbers.go
@@ -2,7 +2,8 @@
 // Use of this source code is governed by a MIT-style license that can be found
 // in the LICENSE file.
 
-package cacher
+// Package numbers provides helper methods to deal with numbers.
+package numbers
 
 import (
 	"bytes"
